+++ include/doubly_linked_list.h
@@ -6,7 +6,7 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/04 15:20:39 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/01/07 16:35:36 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/01/07 16:51:48 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,10 @@ typedef struct e_dll
 	void					(*free)(t_dll **self);
 }							t_dll;
 
+// functions
 t_dll						*init_dll(void);
+
+// methods
 void						add_dll(t_dll *self, int value);
 t_dll_node					*pop_dll(t_dll *self);
 void						swap_dll(t_dll *self);
